pixelator_core::common::node_partitioning

Trait NodePartitioning

pub trait NodePartitioning {

    // Required methods
    fn initialize_from_partitions(partitions: Vec<PartitionId>) -> Self;
    fn get_node_to_partition_map(&self) -> &[PartitionId] ;
    fn update_partition(&mut self, node: NodeIdx, partition: PartitionId);
    fn normalize(&mut self);

    // Provided methods
    fn initialize_with_singlet_partitions(num_nodes: NodeIdx) -> Self
       where Self: Sized { ... }
    fn initialize_with_null_partition(num_nodes: NodeIdx) -> Self
       where Self: Sized { ... }
    fn num_nodes(&self) -> usize { ... }
    fn get_partition_for_node(&self, node: &NodeIdx) -> PartitionId { ... }
    fn get_partition_to_node_map(
        &self,
    ) -> HashMap<PartitionId, HashSet<NodeIdx>> { ... }
    fn num_partitions(&self) -> usize { ... }
    fn is_normalized(&self) -> bool { ... }
    fn into_node_partitioning<T: NodePartitioning>(self) -> T
       where Self: Sized { ... }
    fn partitions_as_matrix(&self) -> CsMat<f64> { ... }
    fn top_k_largest_partitions(&self, k: usize) -> Vec<HashSet<NodeIdx>>  { ... }
    fn largest_partition_size(&self) -> usize { ... }

}

Required Methods

fn initialize_from_partitions(partitions: Vec<PartitionId>) -> Self

Initialize partitions from a vector mapping node ids to partition ids

fn get_node_to_partition_map(&self) -> &[PartitionId]

Returns a reference to the vector mapping node ids to partition ids

fn update_partition(&mut self, node: NodeIdx, partition: PartitionId)

Update the partition of a specific node

fn normalize(&mut self)

Normalizes partition ids to use all integers from 0 to |P| - 1

Provided Methods

fn initialize_with_singlet_partitions(num_nodes: NodeIdx) -> Self
where Self: Sized,

Put all nodes in their own individual partitions

fn initialize_with_null_partition(num_nodes: NodeIdx) -> Self
where Self: Sized,

Put all nodes in the same unique partition

fn num_nodes(&self) -> usize

Returns the total number of nodes

fn get_partition_for_node(&self, node: &NodeIdx) -> PartitionId

Returns the partition id of a given node

fn get_partition_to_node_map(&self) -> HashMap<PartitionId, HashSet<NodeIdx>>

Generate a map from partition id to node ids

fn num_partitions(&self) -> usize

Return the total number of partitions

fn is_normalized(&self) -> bool

Returns true if all partitions ids from 0 to |P| - 1 are used

fn into_node_partitioning<T: NodePartitioning>(self) -> T
where Self: Sized,

Converts current NodePartitioning into another NodePartitioning implementation.

For instance, this is useful to convert a FastNodePartitioning instance into a LeidenNodePartitioning instance during the transition from FLP to Leiden.

fn partitions_as_matrix(&self) -> CsMat<f64>

Generates the |P| x |N| sparse matrix matching nodes to partitions

fn top_k_largest_partitions(&self, k: usize) -> Vec<HashSet<NodeIdx>>

Returns the largest k partitions

fn largest_partition_size(&self) -> usize

Returns the size of the largest partition

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.